+++ src/core/vis/processors/image/depthmask.cpp
@@ -48,7 +48,7 @@ const std::string DepthMask::getProcessorInfo() const {
 }
 
 void DepthMask::process(LocalPortMapping* portMapping) {
-    glViewport(0,0,(int)size_.x,(int)size_.y);
+    glViewport(0,0,size_.x,size_.y);
 
 	int source = portMapping->getTarget("image.inport");
     int dest = portMapping->getTarget("image.outport");
@@ -72,8 +72,8 @@ void DepthMask::process(LocalPortMapping* portMapping) {
     // initialize shader
     program_->activate();
     setGlobalShaderParameters(program_);
-    program_->setUniform("shadeTex_", (GLint) tm_.getTexUnit(shadeTexUnit_));
-    program_->setUniform("depthTex_", (GLint) tm_.getTexUnit(depthTexUnit_));
+    program_->setUniform("shadeTex_", tm_.getTexUnit(shadeTexUnit_));
+    program_->setUniform("depthTex_", tm_.getTexUnit(depthTexUnit_));
     program_->setUniform("lambda_", lambda_.get());
     program_->setUniform("minDepth_", minDepth_.get());
     program_->setUniform("maxDepth_", maxDepth_.get());
@@ -89,6 +89,4 @@ void DepthMask::setLambda(float lambda) {
     lambda_.set(lambda);
 }
 
-
 } // voreen namespace
-
